cleanup: separate util functions by domain
