+++ src/commands/guild-management/addMember.ts
@@ -1,5 +1,5 @@
 import { GuildService } from "@/lib/services/GuildService";
-import { isValidUUIDv4 } from "@/lib/utils";
+import { isValidUUIDv4 } from "@/lib/utils/mathUtilts";
 import {
     ChatInputCommandInteraction,
     MessageFlags,
