+++ src/commands/guild-raid/inactivityBySeason.ts
@@ -4,7 +4,7 @@ import {
     MINIMUM_SEASON_THRESHOLD,
 } from "@/lib/configs/constants";
 import { GuildService } from "@/lib/services/GuildService.ts";
-import { sortTokensUsed } from "@/lib/utils";
+import { sortTokensUsed } from "@/lib/utils/mathUtilts";
 import { Rarity } from "@/models/enums";
 import type { TokensUsed } from "@/models/types/TokensUsed";
 import {
