+++ src/commands/guild-raid/seasonByRarity.ts
@@ -5,11 +5,9 @@ import {
 } from "@/lib/configs/constants";
 import { ChartService } from "@/lib/services/ChartService";
 import { GuildService } from "@/lib/services/GuildService.ts";
-import {
-    numericAverage,
-    numericMedian,
-    sortGuildRaidResultDesc,
-} from "@/lib/utils";
+import { numericMedian } from "@/lib/utils/mathUtilts";
+import { numericAverage } from "@/lib/utils/mathUtilts";
+import { sortGuildRaidResultDesc } from "@/lib/utils/mathUtilts";
 import { Rarity } from "@/models/enums";
 import {
     AttachmentBuilder,
