+++ src/commands/guild-raid/seasonTokens.ts
@@ -5,7 +5,9 @@ import {
 } from "@/lib/configs/constants";
 import { ChartService } from "@/lib/services/ChartService";
 import { GuildService } from "@/lib/services/GuildService.ts";
-import { numericAverage, numericMedian, standardDeviation } from "@/lib/utils";
+import { numericMedian } from "@/lib/utils/mathUtilts";
+import { numericAverage } from "@/lib/utils/mathUtilts";
+import { standardDeviation } from "@/lib/utils/mathUtilts";
 import { Rarity } from "@/models/enums";
 import {
     AttachmentBuilder,
