+++ src/commands/guild-raid/trackMember.ts
@@ -4,7 +4,8 @@ import {
     MAXIMUM_TOKENS_PER_SEASON,
 } from "@/lib/configs/constants";
 import { GuildService } from "@/lib/services/GuildService";
-import { numericAverage, numericMedian } from "@/lib/utils";
+import { numericMedian } from "@/lib/utils/mathUtilts";
+import { numericAverage } from "@/lib/utils/mathUtilts";
 import { Rarity } from "@/models/enums";
 import {
     ChatInputCommandInteraction,
