+++ src/lib/services/GuildService.ts
@@ -1,30 +1,30 @@
 import { HominaTacticusClient } from "@/client";
 import { dbController, logger } from "@/lib";
+import { SecondsToString } from "../utils/timeUtilts";
+import { evaluateToken } from "../utils/timeUtilts";
+import { getUnixTimestamp } from "../utils/timeUtilts";
+import {
+    getMetaTeams,
+    hasLynchpinHeroes,
+    inTeamsCheck,
+} from "@/lib/utils/metaTeamUtils";
+import { DamageType, EncounterType, Rarity } from "@/models/enums";
+import { MetaTeams } from "@/models/enums/MetaTeams";
 import type {
     GuildRaidAvailable,
     GuildRaidResult,
     Raid,
-    TokenStatus,
     TokensAndBombs,
+    TokenStatus,
 } from "@/models/types";
-import { DamageType, EncounterType, Rarity } from "@/models/enums";
-import type { TeamDistribution } from "@/models/types/TeamDistribution";
-import {
-    evaluateToken,
-    getMetaTeams,
-    getUnixTimestamp,
-    hasLynchpinHeroes,
-    inTeamsCheck,
-    SecondsToString,
-    testApiToken,
-} from "../utils";
 import type { GuildMemberMapping } from "@/models/types/GuildMemberMapping";
+import type { MetaComps } from "@/models/types/MetaComps";
+import type { TeamDistribution } from "@/models/types/TeamDistribution";
 import {
     META_TEAM_THRESHOLD,
     MINIMUM_SEASON_THRESHOLD,
 } from "../configs/constants";
-import type { MetaComps } from "@/models/types/MetaComps";
-import { MetaTeams } from "@/models/enums/MetaTeams";
+import { testApiToken } from "../utils/commandUtils";
 
 /**
  * Service class for managing guild-related operations in the Homina Tacticus application.
