+++ src/lib/utils/colorUtils.ts
@@ -1,0 +1,29 @@
+//**
+// ChartJs utils
+//  */
+
+export const CHART_COLORS = {
+    red: "rgb(255, 99, 132)",
+    orange: "rgb(255, 159, 64)",
+    yellow: "rgb(255, 205, 86)",
+    green: "rgb(75, 192, 192)",
+    blue: "rgb(54, 162, 235)",
+    purple: "rgb(153, 102, 255)",
+    grey: "rgb(201, 203, 207)",
+    discordbg: "rgb(57,58,65)",
+};
+
+const NAMED_COLORS = [
+    CHART_COLORS.red,
+    CHART_COLORS.orange,
+    CHART_COLORS.yellow,
+    CHART_COLORS.green,
+    CHART_COLORS.blue,
+    CHART_COLORS.purple,
+    CHART_COLORS.grey,
+    CHART_COLORS.discordbg,
+];
+
+export function namedColor(index: number) {
+    return NAMED_COLORS[index % NAMED_COLORS.length];
+}
