+++ src/lib/utils/mathUtilts.ts
@@ -1,0 +1,49 @@
+import type { GuildRaidResult } from "@/models/types";
+import type { TokensUsed } from "@/models/types/TokensUsed";
+
+// Nb! Relies on the user providing sorted data
+
+export function getTopNDamageDealers(sortedData: GuildRaidResult[], n: number) {
+    return sortedData.slice(0, n).map((player, index) => {
+        const medal = index === 0 ? "🥇" : index === 1 ? "🥈" : "🥉";
+        const formattedDamage = player.totalDamage.toLocaleString();
+        return `${medal} ${player.username}: ${formattedDamage}`;
+    });
+}
+export function sortGuildRaidResultDesc(data: GuildRaidResult[]) {
+    return data.sort((a, b) => b.totalDamage - a.totalDamage);
+}
+export function sortTokensUsed(data: TokensUsed[]) {
+    return data.sort((a, b) => b.tokens - a.tokens);
+}
+export function isValidUUIDv4(uuid: string): boolean {
+    const uuidRegex =
+        /^[0-9a-f]{8}-[0-9a-f]{4}-4[0-9a-f]{3}-[89ab][0-9a-f]{3}-[0-9a-f]{12}$/i;
+    return uuidRegex.test(uuid);
+}
+export function standardDeviation(arr: number[]): number {
+    if (arr.length === 0) return 0;
+
+    const mean = numericAverage(arr);
+    const squaredDiffs = arr.map((value) => Math.pow(value - mean, 2));
+    const variance = numericAverage(squaredDiffs);
+    return Math.sqrt(variance);
+}
+export function numericAverage(arr: number[]): number {
+    if (arr.length === 0) return 0;
+
+    const sum = arr.reduce((acc, val) => acc + val, 0);
+    return sum / arr.length;
+}
+export function numericMedian(arr: number[]): number {
+    if (arr.length === 0) return 0;
+
+    const sorted = [...arr].sort((a, b) => a - b);
+    const middleIndex = Math.floor(sorted.length / 2);
+
+    if (sorted.length % 2 === 0) {
+        return (sorted[middleIndex - 1]! + sorted[middleIndex]!) / 2;
+    } else {
+        return sorted[middleIndex]!;
+    }
+}
